chore: refactor file writer / circuit builder
